fix(frontend): wrap auth callback in suspense for production build
